requirements.txt
python-dotenv==1.0.1
pydantic>=2,<3

# LangChain 0.3.x
langchain==0.3.7
langchain-core==0.3.63
langchain-community==0.3.7
langchain-ollama==0.2.0
langchain-google-genai==2.0.4

# Vector DB / PDF
chromadb==0.5.5
pypdf==4.3.1

# Hugging Face
transformers==4.44.2
huggingface_hub==0.24.6
sentence-transformers==3.0.1

# Audio / Whisper
openai-whisper==20231117
torch==2.2.2

# Misc
tiktoken==0.7.0
numpy==1.26.4
packaging==24.1
urllib3==2.3.0
scikit-learn==1.4.2
scipy==1.11.4